Check linters errors
