+++ index.css
@@ -1,186 +1,186 @@
 nav {
- display: flex;
- justify-content: space-between;
- border: 1px solid black;
- padding: 10px;
+  display: flex;
+  justify-content: space-between;
+  border: 1px solid black;
+  padding: 10px;
 }
 
 .header {
- display: flex;
- align-items: center;
+  display: flex;
+  align-items: center;
 }
 
 a {
- text-decoration: none;
- color: black;
- width: 50px;
+  text-decoration: none;
+  color: black;
+  width: 50px;
 }
 
 ul {
- padding: 0;
- margin: 0;
+  padding: 0;
+  margin: 0;
 }
 
 h4 {
- margin: 0;
+  margin: 0;
 }
 
 .black {
- border-right: 1px solid black;
- border-left: 1px solid black;
- margin: 0 10px;
- padding: 0 5px;
+  border-right: 1px solid black;
+  border-left: 1px solid black;
+  margin: 0 10px;
+  padding: 0 5px;
 }
 
 #time {
- text-align: end;
- margin-top: 10px;
- font-weight: bolder;
+  text-align: end;
+  margin-top: 10px;
+  font-weight: bolder;
 }
 
 .main-section {
- display: flex;
- flex-direction: column;
- margin: auto;
- justify-content: center;
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+  justify-content: center;
 }
 
 h1 {
- color: black;
- text-align: center;
+  color: black;
+  text-align: center;
 }
 
 .container {
- border: 2px;
- border-style: solid;
- border-color: black;
- margin-bottom: 4%;
- border-radius: 3px;
+  border: 2px;
+  border-style: solid;
+  border-color: black;
+  margin-bottom: 4%;
+  border-radius: 3px;
 }
 
 .book {
- display: flex;
- justify-content: space-between;
- border: 1px;
- border-color: black;
- margin: 0;
- padding: 0;
- height: 30px;
+  display: flex;
+  justify-content: space-between;
+  border: 1px;
+  border-color: black;
+  margin: 0;
+  padding: 10px;
+  height: 30px;
 }
 
 .book:nth-child(odd) {
- background-color: rgb(173, 173, 173);
+  background-color: rgb(173, 173, 173);
 }
 
 .book:nth-child(even) {
- background-color: white;
+  background-color: white;
 }
 
 h2 {
- color: black;
- text-align: center;
+  color: black;
+  text-align: center;
 }
 
 li {
- list-style: none;
+  list-style: none;
 }
 
 .contact {
- margin: 30px;
+  margin: 30px;
 }
 
 #contact li {
- list-style: disc;
+  list-style: disc;
 }
 
 form {
- display: flex;
- flex-direction: column;
- align-items: end;
- width: 60%;
- margin: auto;
+  display: flex;
+  flex-direction: column;
+  align-items: end;
+  width: 60%;
+  margin: auto;
 }
 
 #book-form {
- display: flex;
- gap: 10px;
+  display: flex;
+  gap: 10px;
 }
 
 .about-border {
- display: block;
- width: 20%;
- height: 2px;
- background: black;
- margin: auto;
+  display: block;
+  width: 20%;
+  height: 2px;
+  background: black;
+  margin: auto;
 }
 
 form button {
- padding: 3px 10px;
- background-color: white;
- box-shadow: 2px 2px;
- border: 2px solid black;
+  padding: 3px 10px;
+  background-color: white;
+  box-shadow: 2px 2px;
+  border: 2px solid black;
 }
 
 .book button {
- align-self: center;
- background-color: white;
- box-shadow: 2px 2px;
- border: 2px solid black;
- font-size: 16px;
- margin-right: 10px;
+  align-self: center;
+  background-color: white;
+  box-shadow: 2px 2px;
+  border: 2px solid black;
+  font-size: 16px;
+  margin-right: 10px;
 }
 
 input {
- width: 98%;
- margin: auto;
- height: 30px;
- border: 2px solid black;
- box-shadow: 1px 1px;
+  width: 98%;
+  margin: auto;
+  height: 30px;
+  border: 2px solid black;
+  box-shadow: 1px 1px;
 }
 
 #contact {
- display: flex;
- flex-direction: column;
- margin: auto;
+  display: flex;
+  flex-direction: column;
+  margin: auto;
 }
 
 footer {
- border: 2px solid black;
- margin-top: 20px;
+  border: 2px solid black;
+  margin-top: 20px;
 }
 
 p {
- margin: 10px;
+  margin: 10px;
 }
 
 @media screen and (min-width: 1020px) {
- body {
-   margin: 20px 150px;
- }
+  body {
+    margin: 20px 150px;
+  }
 
- .main-section {
-   border: 2px solid black;
-   padding: 10px;
-   width: 60%;
-   margin: auto;
- }
+  .main-section {
+    border: 2px solid black;
+    padding: 10px;
+    width: 60%;
+    margin: auto;
+  }
 
- #contact {
-   display: flex;
-   flex-direction: column;
-   margin: auto;
- }
+  #contact {
+    display: flex;
+    flex-direction: column;
+    margin: auto;
+  }
 
- footer {
-   border: 2px solid black;
-   padding: 0 10px;
-   margin-top: 20px;
- }
+  footer {
+    border: 2px solid black;
+    padding: 0 10px;
+    margin-top: 20px;
+  }
 }
 
 @media screen and (min-width: 1440px) {
- .main-section {
-   width: 50%;
-   margin: auto;
- }
+  .main-section {
+    width: 50%;
+    margin: auto;
+  }
 }
