+++ index.js
@@ -1,37 +1,24 @@
-import Book from '/modules/storeBook.js';
-import { getBookList } from '/module/bookList.js'
-import { setupRemove } from '/module/removeBooks.js'
-import { addBook } from '/module/addBooks.js'
-
-class Books {
-    constructor() {
-        this.books = [];
-    }
-    initialize() {
-        const dataString = localStorage.getItem('bookData');
-        if (dataString) {
-            this.books = JSON.parse(dataString).map((book) => new Book(book.title, book.author));
-            this.setHtml();
-        }
-        setupRemove();
-    }
-    newBook(title, author) {
-        const book = new Book(title, author);
-        this.books.push(book);
-        localStorage.setItem('bookData', JSON.stringify(this.books));
-        this.setHtml();
-        return book;
-    }
-
-    setHtml() {
-        const container = document.querySelector('.container');
-        container.innerHTML = getBookList();
-        setupRemove();
-    }
-}
+import Books from './modules/removeBooks.js';
+import spa from './modules/spa.js';
 
 const bookRepo = new Books();
 
 bookRepo.initialize();
 
+const addBook = () => {
+  const form = document.getElementById('book-form');
+  form.addEventListener('submit', (e) => {
+    e.preventDefault();
+    const { title, author } = form.elements;
+    bookRepo.newBook(title.value, author.value);
+
+    title.value = '';
+    author.value = '';
+  });
+};
+
 addBook();
+spa();
+
+const d = new Date();
+document.getElementById('time').innerHTML = d.toUTCString();
