--- modules/addDate.js
@@ -1,2 +1,0 @@
-const d = new Date();
-document.getElementById('time').innerHTML = d.toUTCString();
