--- modules/bookList.js
@@ -1,9 +1,0 @@
-const getBookList = () => {
-    let containerHtml = '';
-    this.books.forEach((book) => {
-        containerHtml += book.getHtml();
-    });
-    return containerHtml;
-}
-
-export default getBookList;
