+++ modules/storeBook.js
@@ -1,14 +1,12 @@
 export default class Book {
-    constructor(title, author) {
-        this.title = title;
-        this.author = author;
-    }
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
 
-    getHtml() {
-        return `<ul class="book">
+    getHtml = () => `<ul class="book">
      <li class="list">"${this.title}" by 
      ${this.author}</li>
      <button class="remove">Remove</button>
-     </ul>`;
-    }
+     </ul>`
 }
